requirements-dev.txt
coverage==4.5.1
Django>=2.0
djangorestframework>=3.8
pycodestyle>=2.3
django-filter>=2.2.0
